requirements.txt
replicate==0.23.1
python-dotenv==1.0.1